+++ Sources/GRPC/ClientCalls/ClientCall.swift
@@ -164,45 +164,3 @@ public protocol UnaryResponseClientCall: ClientCall {
   /// Callers should rely on the `status` of the call for the canonical outcome.
   var response: EventLoopFuture<ResponsePayload> { get }
 }
-
-// These protocols sit between a `*ClientCall` and the channel. The intention behind them is to
-// allow the `*ClientCall` types to use a different transport mechanisms. Normally this will be
-// a NIO HTTP/2 stream channel.
-
-internal protocol ClientCallInbound {
-  associatedtype Response
-  typealias ResponsePart = _GRPCClientResponsePart<Response>
-
-  /// Receive an error.
-  func receiveError(_ error: Error)
-
-  /// Receive a response part.
-  func receiveResponse(_ part: ResponsePart)
-}
-
-internal protocol ClientCallOutbound {
-  associatedtype Request
-  typealias RequestPart = _GRPCClientRequestPart<Request>
-
-  /// Send a single request part and complete the promise once the part has been sent.
-  func sendRequest(_ part: RequestPart, promise: EventLoopPromise<Void>?)
-
-  /// Send the given request parts and complete the promise once all parts have been sent.
-  func sendRequests<S: Sequence>(_ parts: S, promise: EventLoopPromise<Void>?)
-    where S.Element == RequestPart
-
-  /// Cancel the call and complete the promise once the cancellation has been completed.
-  func cancel(promise: EventLoopPromise<Void>?)
-}
-
-extension ClientCallOutbound {
-  /// Convenience method for sending all parts of a unary-request RPC.
-  internal func sendUnary(_ head: _GRPCRequestHead, request: Request, compressed: Bool) {
-    let parts: [RequestPart] = [
-      .head(head),
-      .message(_MessageContext(request, compressed: compressed)),
-      .end,
-    ]
-    self.sendRequests(parts, promise: nil)
-  }
-}
